requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
nats-py
pydantic
python-dotenv